+++ app/app/Http/Controllers/FormComponentsController.php
@@ -116,7 +116,7 @@ class FormComponentsController
     public function libraryDefaults()
     {
         return view('form.components.libraryDefaults', [
-            'defaults'  => [
+            'defaults' => [
                 'biography' => 'Voluptate ea culpa proident proident qui nostrud non ea irure ullamco in non reprehenderit.',
                 'country'   => 'NL',
                 'countries' => ['BE', 'NL'],
@@ -147,7 +147,7 @@ class FormComponentsController
     public function libraryChange()
     {
         return view('form.components.libraryChange', [
-            'defaults'  => [
+            'defaults' => [
                 'biography' => 'Voluptate ea culpa proident proident qui nostrud non ea irure ullamco in non reprehenderit.',
                 'country'   => 'NL',
                 'countries' => ['BE', 'NL'],
@@ -217,7 +217,8 @@ class FormComponentsController
 
     public function submit(Request $request)
     {
-        $request->validate([
+        // Validate with another bag to make sure this also works.
+        $request->validateWithBag('dummy', [
             'name'      => ['required', 'string'],
             'password'  => ['required', 'string'],
             'secret'    => ['required', 'string'],
@@ -241,15 +242,15 @@ class FormComponentsController
         $countries = array_keys($this->countries());
 
         $request->validate([
-            'country_a'     => ['required', 'string', Rule::in($countries)],
-            'country_b'     => ['required', 'string', Rule::in($countries)],
-            'country_c'     => ['required', 'string', Rule::in($countries)],
-            'country_d'     => ['required', 'string', Rule::in($countries)],
+            'country_a' => ['required', 'string', Rule::in($countries)],
+            'country_b' => ['required', 'string', Rule::in($countries)],
+            'country_c' => ['required', 'string', Rule::in($countries)],
+            'country_d' => ['required', 'string', Rule::in($countries)],
 
-            'countries_a'   => ['required', 'array', 'min:1'],
-            'countries_b'   => ['required', 'array', 'min:1'],
-            'countries_c'   => ['required', 'array', 'min:1'],
-            'countries_d'   => ['required', 'array', 'min:1'],
+            'countries_a' => ['required', 'array', 'min:1'],
+            'countries_b' => ['required', 'array', 'min:1'],
+            'countries_c' => ['required', 'array', 'min:1'],
+            'countries_d' => ['required', 'array', 'min:1'],
 
             'countries_a.*' => ['required', 'string', Rule::in($countries)],
             'countries_b.*' => ['required', 'string', Rule::in($countries)],