+++ app/app/Http/Controllers/ModalController.php
@@ -37,6 +37,11 @@ class ModalController
         return view('modal.validation');
     }
 
+    public function opened()
+    {
+        return view('modal.opened');
+    }
+
     public function size($size)
     {
         return view('modal.size', ['size' => $size]);