+++ app/config/translator.php
@@ -11,9 +11,9 @@ return [
         base_path('../resources/views'),
         base_path('../src'),
     ],
-    'output'      => base_path('../resources/lang'),
-    'extensions'  => ['php'],
-    'container'   => [
+    'output'     => base_path('../resources/lang'),
+    'extensions' => ['php'],
+    'container'  => [
         'config_loader'          => LaravelConfigLoader::class,
         'translation_repository' => LaravelJsonTranslationRepository::class,
     ],