+++ app/resources/views/form/redirect.blade.php
@@ -1,0 +1,11 @@
+@extends('layout')
+
+@section('content')
+
+FormRedirect
+
+<x-splade-form class="form-simple">
+    <button type="submit">Submit</button>
+</x-splade-form>
+
+@endsection