+++ resources/views/functional/data.blade.php
@@ -1,6 +1,8 @@
 <SpladeData {{ $attributes }}
-    @if($data) :default="@js($data)" @else :default="{!! $json !!}" @endif>
-    <template #default="{!! $scope !!}">
-        {{ $slot }}
-    </template>
+    @if($data) :default="@js($data)" @else :default="{!! $json !!}" @endif :remember="@js($remember)" :local-storage="@js($localStorage)">
+    @unless($store)
+        <template #default="{!! $scope !!}">
+            {{ $slot }}
+        </template>
+    @endunless
 </SpladeData>