+++ src/Http/BladeDirectives.php
@@ -19,6 +19,7 @@ class BladeDirectives
         Blade::directive('spladeHead', [$this, 'spladeHead']);
         Blade::directive('preserveScroll', [$this, 'preserveScroll']);
 
+        $this->registerSEODirectives();
         $this->registerTableCellDirective();
     }
 
@@ -88,6 +89,36 @@ class BladeDirectives
         $function .= " use ({$slotUses})";
 
         return [$name, $function];
+    }
+
+    /**
+     * Registers the Blade SEO directives.
+     *
+     * @return void
+     */
+    public function registerSEODirectives()
+    {
+        $title       = config('splade.blade.seo_title_directive', 'seoTitle');
+        $description = config('splade.blade.seo_description_directive', 'seoDescription');
+        $keywords    = config('splade.blade.seo_keywords_directive', 'seoKeywords');
+
+        if ($title !== false) {
+            Blade::directive($title, function ($expression) {
+                return "<?php \ProtoneMedia\Splade\Facades\SEO::title($expression); ?>";
+            });
+        }
+
+        if ($description !== false) {
+            Blade::directive($description, function ($expression) {
+                return "<?php \ProtoneMedia\Splade\Facades\SEO::description($expression); ?>";
+            });
+        }
+
+        if ($keywords !== false) {
+            Blade::directive($keywords, function ($expression) {
+                return "<?php \ProtoneMedia\Splade\Facades\SEO::keywords($expression); ?>";
+            });
+        }
     }
 
     /**