+++ src/Http/ConfirmPasswordController.php
@@ -4,10 +4,29 @@ namespace ProtoneMedia\Splade\Http;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Validation\ValidationException;
 use ProtoneMedia\Splade\PasswordValidator;
 
 class ConfirmPasswordController
 {
+    /**
+     * Indicates whether the user has recently confirmed their password.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \ProtoneMedia\Splade\PasswordValidator  $passwordValidator
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Request $request, PasswordValidator $passwordValidator): Response
+    {
+        if ($passwordValidator->recentlyConfirmed($request)) {
+            return response()->noContent(200)->skipSpladeMiddleware();
+        }
+
+        throw ValidationException::withMessages([
+            'password' => __('The password confirmation has expired.'),
+        ]);
+    }
+
     /**
      * Confirm the given user's password.
      *
@@ -15,7 +34,7 @@ class ConfirmPasswordController
      * @param  \ProtoneMedia\Splade\PasswordValidator  $passwordValidator
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request, PasswordValidator $passwordValidator): Response
+    public function store(Request $request, PasswordValidator $passwordValidator): Response
     {
         $passwordValidator->validateRequest($request, 'password');