+++ src/Http/SpladeMiddleware.php
@@ -8,8 +8,11 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response as LaravelResponse;
 use Illuminate\Routing\UrlGenerator;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Blade;
+use Illuminate\Support\MessageBag;
 use Illuminate\Support\Str;
+use Illuminate\Support\ViewErrorBag;
 use Illuminate\Validation\ValidationException;
 use Illuminate\View\View;
 use ProtoneMedia\Splade\Facades\Splade;
@@ -22,6 +25,11 @@ use Symfony\Component\HttpFoundation\StreamedResponse;
 
 class SpladeMiddleware
 {
+    /**
+     * @var callable|null
+     */
+    private static $afterOriginalResponseCallback = [];
+
     const FLASH_TOASTS = 'splade.flashToasts';
 
     /**
@@ -47,12 +55,24 @@ class SpladeMiddleware
     {
         // Set and restore some defaults before handling the request.
         $this->splade->setModalKey(Str::uuid());
+        $this->splade->resetDataStores();
         $this->splade->resetLazyComponentCounter();
         $this->splade->resetRehydrateComponentCounter();
         $this->splade->resetPersistentLayoutKey();
 
+        /** @var Session */
+        $session = session()->driver();
+
+        $errorsFromRedirect = $session->pull('errors', new ViewErrorBag);
+
         /** @var Response $response */
         $response = $next($request);
+
+        if (static::$afterOriginalResponseCallback) {
+            $callback = static::$afterOriginalResponseCallback;
+            $callback($request, $response);
+        }
+
         $response->headers->add(['Vary' => 'X-Splade']);
 
         // Don't mess with file and streamed responses.
@@ -68,7 +88,7 @@ class SpladeMiddleware
         }
 
         // Gather the required meta data for the app.
-        $spladeData = $this->spladeData($request->session());
+        $spladeData = $this->spladeData($session, $errorsFromRedirect);
 
         // The response should redirect away from the Splade app.
         if ($redirect = $this->shouldRedirectsAway($response)) {
@@ -78,7 +98,7 @@ class SpladeMiddleware
         // If the response is a redirect, put the toasts into the session
         // so they won't be lost in the next request.
         if (in_array($response->getStatusCode(), [302, 303])) {
-            $request->session()->put(static::FLASH_TOASTS, $this->splade->getToasts());
+            $session->put(static::FLASH_TOASTS, $this->splade->getToasts());
         }
 
         // A Splade request is a request made by the Vue app, so not the initial first request.
@@ -99,14 +119,26 @@ class SpladeMiddleware
     }
 
     /**
+     * A callback that sits between the 'regular' response, and
+     * the handling of the Splade response.
+     *
+     * @param  Closure|null  $callback
+     * @return void
+     */
+    public static function afterOriginalResponse(Closure $callback = null)
+    {
+        static::$afterOriginalResponseCallback = $callback;
+    }
+
+    /**
      * Handle a Splade request, made from the Vue app.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \Illuminate\Http\Response  $response
      * @param  object  $spladeData
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\Response|\Illuminate\Http\JsonResponse
      */
-    private function handleSpladeRequest(Request $request, Response $response, object $spladeData): Response
+    private function handleSpladeRequest(Request $request, Response $response, object $spladeData): Response|JsonResponse
     {
         // We don't mess with JsonResponses, except we add the Splade data to it.
         if ($response instanceof JsonResponse) {
@@ -123,10 +155,14 @@ class SpladeMiddleware
 
             // Get the Validation Errors from the exception and put them in the Splade data.
             if ($response->exception instanceof ValidationException) {
-                $newData['splade']->errors = $response->exception->errors();
+                $newData['splade']->errors = (object) $response->exception->errors();
             }
 
             return $response->setData($newData);
+        }
+
+        if ($response->isRedirect() && $this->splade->dontRefreshPage()) {
+            $response = response()->noContent(200);
         }
 
         if (!$response->isSuccessful()) {
@@ -136,13 +172,16 @@ class SpladeMiddleware
         // Get the rendered content...
         $content = $response->getContent() ?: '';
 
+        // When there are Data Stores registered, wrap the content...
+        $content = $this->wrapContentInDataStores($content);
+
         // Extract the Dynamic Content, we'll return that separately so Vue can handle it.
         [$content, $dynamics] = static::extractDynamicsFromContent($content);
 
         if ($this->splade->isLazyRequest()) {
-            $content = PrepareViewWithLazyComponents::extractComponent($content, $this->splade->getLazyComponentKey()) ?: $content;
+            $content = static::extractComponent($content, 'lazy', $this->splade->getLazyComponentKey()) ?: $content;
         } elseif ($this->splade->isRehydrateRequest()) {
-            $content = PrepareViewWithRehydrateComponents::extractComponent($content, $this->splade->getRehydrateComponentKey());
+            $content = static::extractComponent($content, 'rehydrate', $this->splade->getRehydrateComponentKey());
         }
 
         return $response->setContent(json_encode([
@@ -151,6 +190,32 @@ class SpladeMiddleware
             'dynamics' => $dynamics,
             'splade'   => $spladeData,
         ]));
+    }
+
+    /**
+     * Grabs the component from the rendered content and returns it.
+     *
+     * @param  string  $content
+     * @param  int  $componentKey
+     * @return string
+     */
+    public static function extractComponent(string $content, string $component, int $componentKey): string
+    {
+        $component = strtoupper($component);
+
+        preg_match_all('/START-SPLADE-' . $component . '-(\w+)-->/', $content, $matches);
+
+        return (string) collect($matches[1] ?? [])
+            ->mapWithKeys(function (string $name) use ($content, $component) {
+                $rehydrate = Str::between(
+                    $content,
+                    "<!--START-SPLADE-{$component}-{$name}-->",
+                    "<!--END-SPLADE-{$component}-{$name}-->"
+                );
+
+                return [$name => trim($rehydrate)];
+            })
+            ->get($componentKey);
     }
 
     /**
@@ -183,8 +248,11 @@ class SpladeMiddleware
             $originalViewData = $response->getOriginalContent()->getData();
         }
 
+        // When there are Data Stores registered, wrap the content...
+        $content = $this->wrapContentInDataStores($originalContent);
+
         // Extract the Dynamic Content, we'll return that separately so Vue can handle it.
-        [$content, $dynamics] = static::extractDynamicsFromContent($originalContent);
+        [$content, $dynamics] = static::extractDynamicsFromContent($content);
 
         $viewData = [
             'components' => static::renderedComponents(),
@@ -232,6 +300,30 @@ class SpladeMiddleware
         $wrappedView->with($originalViewData);
 
         return $response;
+    }
+
+    /**
+     * When there are Data Stores registered, wrap the content in a Data Store component.
+     *
+     * @param  string  $content
+     * @return string
+     */
+    private function wrapContentInDataStores(string $content): string
+    {
+        if (empty($this->splade->getDataStores())) {
+            return $content;
+        }
+
+        return Blade::render(implode('', [
+            '<div>',    // We have to start with a div, otherwise the applied backdrop styling in SpladeApp.vue won't work.
+            '<x-splade-component is="data-stores" :stores="$stores">',
+            '{!! $originalContent !!}',
+            '</x-splade-component>',
+            '</div>',
+        ]), [
+            'stores'          => $this->splade->getDataStores(),
+            'originalContent' => $content,
+        ]);
     }
 
     /**
@@ -285,12 +377,49 @@ class SpladeMiddleware
     }
 
     /**
+     * Returns all error messages from the session.
+     *
+     * @param  \Illuminate\Support\ViewErrorBag  $viewErrorBag
+     * @return array
+     */
+    private function allErrorMessages(ViewErrorBag $viewErrorBag): array
+    {
+        return collect($viewErrorBag->getBags())
+            ->flatMap->getMessages()
+            ->toArray();
+    }
+
+    /**
+     * Merges all bags from all view errors bags into one.
+     *
+     * @param  \Illuminate\Support\ViewErrorBag[]  ...$viewErrorsBags
+     * @return \Illuminate\Support\ViewErrorBag
+     */
+    private function mergeViewErrorBags(...$viewErrorsBags): ViewErrorBag
+    {
+        $mergedViewBag = new ViewErrorBag;
+
+        collect($viewErrorsBags)->each(function (ViewErrorBag $viewErrorBag) use ($mergedViewBag) {
+            collect($viewErrorBag->getBags())->each(function (MessageBag $bag, string $key) use ($mergedViewBag) {
+                $mergedBag = $mergedViewBag->hasBag($key)
+                    ? $mergedViewBag->getBag($key)
+                    : tap(new MessageBag, fn ($bag) => $mergedViewBag->put($key, $bag));
+
+                $mergedBag->merge($bag);
+            });
+        });
+
+        return $mergedViewBag;
+    }
+
+    /**
      * This methods returns all relevant data for a Splade page view.
      *
      * @param  \Illuminate\Contracts\Session\Session  $session
+     * @param  \Illuminate\Support\ViewErrorBag  $errorsFromRedirect
      * @return object
      */
-    private function spladeData(Session $session): object
+    private function spladeData(Session $session, ViewErrorBag $errorsFromRedirect): object
     {
         $flashData = config('splade.share_session_flash_data')
             ? collect($session->get('_flash.old', []))
@@ -302,14 +431,19 @@ class SpladeMiddleware
 
         $excludeHead = $this->splade->isLazyRequest() || $this->splade->isRehydrateRequest();
 
+        $mergedViewErrorBag = $this->mergeViewErrorBags(
+            $session->get('errors', new ViewErrorBag),
+            $errorsFromRedirect
+        );
+
         return (object) [
-            'head'             => $excludeHead ? [] : $this->splade->head()->toArray(),
-            'modal'            => $this->splade->isModalRequest() ? $this->splade->getModalType() : null,
-            'modalTarget'      => $this->splade->getModalTarget() ?: null,
-            'flash'            => (object) $flash,
-            'errors'           => (object) session('errors')?->toArray(),
-            'shared'           => (object) $this->splade->getShared(),
-            'toasts'           => array_merge(
+            'head'        => $excludeHead ? [] : $this->splade->head()->toArray(),
+            'modal'       => $this->splade->isModalRequest() ? $this->splade->getModalType() : null,
+            'modalTarget' => $this->splade->getModalTarget() ?: null,
+            'flash'       => (object) $flash,
+            'errors'      => (object) $this->allErrorMessages($mergedViewErrorBag),
+            'shared'      => (object) Arr::map($this->splade->getShared(), fn ($value) => value($value)),
+            'toasts'      => array_merge(
                 $session->pull(static::FLASH_TOASTS, []),
                 $this->splade->getToasts(),
             ),